+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Error during large filter list compiling
+
+[Unreleased]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.127...HEAD
+
+
 ## [v1.1.127] - 2024-03-29
 
 ### Changed